Extract functions to their own class #1
